+++ R/onload.R
@@ -22,6 +22,7 @@
     "https://clinicaltrials.gov/ct2/about-site/terms-conditions#Use\n",
     "https://www.isrctn.com/page/faqs#usingISRCTN\n",
     "https://euclinicaltrials.eu/about-this-website/\n",
+    "\nPlease cite this package, see citation(\"ctrdata\").\n",
     "\n", vTxt[grepl("^Version", vTxt)]
   )