+++ NEWS.md
@@ -1,7 +1,6 @@
-# ctrdata 1.13.9000 (2023-05-01)
- - new development version
+# ctrdata 1.13.9000 (2023-05-07)
  - load more CTIS data (publicEvaluation) and download documents
- - integrate documents downloading into ctrLoadQueryIntoDb() also for CTGOV
+ - integrate downloading documents into ctrLoadQueryIntoDb() also for CTGOV
  - use documents.path for CTGOV, EUCTR, CTIS; deprecated euctrresultsfilespath
  - added documents.regexp to select documents for downloading based on their file name